Revert swid-maven-plugin dependency to public version
The dependency listed here references a version that does not exist
publicly.  It's possible this was a global find/replace error.

I've confirmed this will build with JDK versions 8 and 11.

With apologies, this introduces an end-of-line character change that I
can't figure out how to address between my attempts to edit on Linux,
Mac, Windows, and finally Github's online editor.